+++ src/BundlerMinifierVsix/Adornments/AdornmentProvider.cs
@@ -105,6 +105,7 @@ namespace BundlerMinifierVsix
                         if (bundle.GetAbsoluteOutputFile().Equals(normalizedFilePath, StringComparison.OrdinalIgnoreCase))
                         {
                             GeneratedAdornment generated = new GeneratedAdornment(textView, _isVisible, _initOpacity);
+                            textView.Properties.AddProperty("generated", true);
                             break;
                         }
                     }