+++ src/BundlerMinifierVsix/BundleService.cs
@@ -102,7 +102,7 @@ namespace BundlerMinifierVsix
                 catch (Exception ex)
                 {
                     Logger.Log(ex);
-                    MessageBox.Show($"There is an error in the {Constants.CONFIG_FILENAME} file. This could be due to a change in the format after this extension was updated.", "Web Compiler", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                    MessageBox.Show($"There is an error in the {Constants.CONFIG_FILENAME} file. This could be due to a change in the format after this extension was updated.", Constants.VSIX_NAME, MessageBoxButtons.OK, MessageBoxIcon.Information);
                 }
             });
         }
@@ -118,7 +118,7 @@ namespace BundlerMinifierVsix
                 catch (Exception ex)
                 {
                     Logger.Log(ex);
-                    MessageBox.Show($"There is an error in the {Constants.CONFIG_FILENAME} file. This could be due to a change in the format after this extension was updated.", "Web Compiler", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                    MessageBox.Show($"There is an error in the {Constants.CONFIG_FILENAME} file. This could be due to a change in the format after this extension was updated.", Constants.VSIX_NAME, MessageBoxButtons.OK, MessageBoxIcon.Information);
                 }
             });
         }